requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.16
django-filter>=25.0
Pillow>=10.0
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.2
